resource: Refactor resource data code into distinct headers/sources
Move schedule data code into sched_data.[h|c]pp.

Move scheduler infrastructure data code into infra_data.[h|c]pp.

Introduce resource_data.cpp to keep the header file clean.
